requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
pydantic==2.3.0
pydantic-settings==2.0.3
email-validator==2.2.0
aio-pika==9.3.0
SQLAlchemy==2.0.20
alembic==1.12.0
asyncpg==0.28.0
anyio==4.6.2
prometheus-client
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1